Remove wind and pressure tolerances
Remove the parameters wind_tolerance and pressure_tolerance as they are
no longer used in the new source term and storm field settings.  These
were replaced by a spatial ramping function based on the distance from
the eye of the storm and the radius of the last closed isobar.  In the
future the ramp_width parameter may be added instead.