Fix data file writing bug
Occasionally when data files were being written out in the same process as
a run the buffer for each of the files were never being actually dumped into
the files themselves leaving emtpy files.  The call to `self.close_data_file`
is added in this commit to all the data objects that are missing it.