Fix track output bug
The track output was moved into tick from valout to reduce the number of
duplicated routines in the surge code.  Unfortunately the initial
location of the storm was no longer being output.  This commit returns
the output of the track to valout, removes it from tick, and changes the
warning in the plotting from an exception to a printed warning.