First pass through reworking multilayer code to 5.0
Commit includes updates to a lot of the core multilayer routines for
compatibility with Clawpack 5.0 release.  Includes sectioning of much
of the actual data parameters to their own modules or data structures
and the reintroduction of some older codes.  Also makes some
modifications to existing surge code so that data structures play
nicely together including some changes to default values.  This also
corresponds to a change to the rundata object in clawutil in order
to be able to replace the default Qinit with the multilayer specfic
one.