Initial commit of topo module containing base topography classes
This includes most of the intended funtionality of a topography class
that is intended to act as a representation of a topography file.  In
particular it will defer calculation of forms of the data until requested
and includes the ability to plot and project unstructured data onto a
structured grid.