Fix seg-fault where synthetic topo was used
Since it is not clear how to deal with synthetic topography with dtopo files
it is left up to the specification of the topography to determine the changing
topography if applicable.  This patch corrects for the fact that certain
parameters were not being set (which caused a seg-fault) and also removes
a deallocation problem.