Add newer way to calculate geometry
Combine the old way to compute the geometry dictionary for subfaults with the
method used in a previous proto-type of the SubFault class.  Adds more
coordinate information and is easier to read.  Also adds a test to compare
the previous implementation with the new.  More work is needed to clean up
where things are expecting the dictionary, currently the new arrays
fault_plane_centers and fault_plane_corners are being unpacked into the
previous variables.