Move initialization routines
This commit moves most of the GeoClaw intialization routines into the
amr2 routine leaving setprob to only contain user changes and not
geoclaw data always expected to be generated.  This also adds to some of
the modules a check to see if the module has been initialized.  This
shoudl be added to all of the standard modules for backwards
compatibility.  Also note now that this inclued the multilayer data by
default.