Simplify storm specification and add more capabilities
The major change is reducing the number of input parameters to
one where the "data" and "model" storm types are only one input
parameter and instead just use integers that are negative and
positive respectively for their specification.  The issue with
the storm fields being to small is still present but right now it
seems to be due to a parameter conversion error that has yet to be
found.