topo_module.f90 fix for grid registration
If topofile specifies xllcorner,yllcorner then shift x,y by dx/2,dy/2.
This is the correct thing to do but differs from previous versions of geoclaw,
which did not handle this properly. So computational results may differ if
topofiles are used that specify this registration in the header, in particular
etopo1 files or other obtained in .asc format from NCEI DEM database.

To revert to earlier behavior, with values from header specifying the location
of the first grid point (rather than lower left corner of cell), change
xllcorner,yllcorner to xlower,ylower.